# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Inventory Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"

    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found

    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /stock/movements:
    get:
      operationId: listStockMovements
      tags: [stock]
      summary: List stock movements
      parameters:
        - name: productId
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Stock movements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockMovement"

    post:
      operationId: createStockMovement
      tags: [stock]
      summary: Create stock movement (in or out)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockMovementInput"
      responses:
        "201":
          description: Created movement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockMovement"
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, categoryId, categoryName, quantity, unit, price, status, barcode, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        categoryId:
          type: integer
        categoryName:
          type: string
        quantity:
          type: integer
        unit:
          type: string
        price:
          type: number
        lowStockThreshold:
          type: integer
        status:
          type: string
        barcode:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        brand:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ProductList:
      type: object
      required: [products, total, page, limit]
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    ProductInput:
      type: object
      required: [name, categoryId, quantity, unit, price]
      properties:
        name:
          type: string
        categoryId:
          type: integer
        quantity:
          type: integer
        unit:
          type: string
        price:
          type: number
        lowStockThreshold:
          type: integer
        barcode:
          type: string
        description:
          type: string
        brand:
          type: string
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        categoryId:
          type: integer
        quantity:
          type: integer
        unit:
          type: string
        price:
          type: number
        lowStockThreshold:
          type: integer
        barcode:
          type: string
        description:
          type: string
        brand:
          type: string
...
    StockMovement:
      type: object
      required: [id, productId, productName, type, quantity, reason, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        type:
          type: string
        quantity:
          type: integer
        reason:
          type: ["string", "null"]
        createdAt:
          type: string
    StockMovementInput:
      type: object
      required: [productId, type, quantity]
      properties:
        productId:
          type: integer
        type:
          type: string
        quantity:
          type: integer
        reason:
          type: string
...