# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Inventory Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  categoryId: number;
  categoryName: string;
  quantity: number;
  unit: string;
  price: number;
  lowStockThreshold?: number;
  status: string;
  /** @nullable */
  barcode: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  brand?: string | null;
  createdAt: string;
  updatedAt: string;
}