# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Inventory Management API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  categoryId: number;
  quantity: number;
  unit: string;
  price: number;
  lowStockThreshold?: number;
  barcode?: string;
  description?: string;
  brand?: string;
}