# lib/api-zod/src/generated/types/stockMovement.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Inventory Management API
 * OpenAPI spec version: 0.1.0
 */

export interface StockMovement {
  id: number;
  productId: number;
  productName: string;
  type: string;
  quantity: number;
  /** @nullable */
  reason: string | null;
  createdAt: string;
}